third_party/WebKit/LayoutTests/fast/loader/crash-focus-in-unload.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<body>
<iframe></iframe>
</body>
<script>
'use strict';

let iframe = document.querySelector('iframe');
let iframeDoc = iframe.contentDocument;

let focusableIframe = document.createElement('iframe');
iframeDoc.body.appendChild(focusableIframe);

focusableIframe.contentWindow.addEventListener('unload', () => {
  // Try to focus on frame owner during unload event of its content.
  focusableIframe.focus();
});

test(() => {
  // Trigger unload event.
  document.body.appendChild(focusableIframe);

  var input = document.createElement('input');
  iframeDoc.body.appendChild(input);
  input.focus();
  assert_true(true, 'This test must run without crash.');
}, 'Should not trigger assertion. crbug.com/677690');
</script>
